Move #validate_status and specs to Command
It is now a Command's responsibility to validate the execution status.
The exceptions are part of its API.
