+++ spec/percona_migrator/runner_spec.rb
@@ -7,7 +7,6 @@ describe PerconaMigrator::Runner do
     instance_double(
       PerconaMigrator::Logger,
       write: true,
-      write_no_newline: true,
       say: true
     )
   end
@@ -15,12 +14,7 @@ describe PerconaMigrator::Runner do
   let(:mysql_adapter) do
     instance_double(ActiveRecord::ConnectionAdapters::Mysql2Adapter)
   end
-  let(:config) do
-    instance_double(
-      PerconaMigrator::Configuration,
-      error_log_path: 'percona_migrator_error.log'
-    )
-  end
+  let(:config) { instance_double(PerconaMigrator::Configuration) }
 
   let(:runner) { described_class.new(logger, cli_generator, mysql_adapter, config) }
 
@@ -41,37 +35,17 @@ describe PerconaMigrator::Runner do
   end
 
   describe '#execute' do
-    let(:temp_file) do
-      file = Tempfile.new('faked_stdout')
-      file.write('hello world\ntodo roto')
-      file.rewind
-      file.close
-      file
-    end
-    let(:status) do
-      instance_double(
-        Process::Status,
-        exitstatus: 0,
-        signaled?: false,
-        success?: true
-      )
-    end
-    let(:stdout) { temp_file.open }
-    let(:wait_thread) { instance_double(Thread, value: status) }
-    let(:expected_command) { "#{command} 2> #{config.error_log_path}" }
+    let(:status) { instance_double(Process::Status) }
+    let(:cmd) { instance_double(PerconaMigrator::Command, run: status) }
 
     before do
-      allow(Open3).to(
-        receive(:popen3)
-        .with(expected_command)
-        .and_yield(nil, stdout, nil, wait_thread)
-      )
+      allow(PerconaMigrator::Command)
+        .to receive(:new).with(command, config, logger).and_return(cmd)
     end
 
     it 'executes the pt-online-schema-change command' do
       runner.execute(command)
-
-      expect(Open3).to have_received(:popen3).with(expected_command)
+      expect(cmd).to have_received(:run)
     end
 
     it 'returns the command status' do
@@ -84,64 +58,6 @@ describe PerconaMigrator::Runner do
         "Running pt-online-schema-change command\n\n",
         true
       )
-    end
-
-    it 'logs the command\'s output' do
-      runner.execute(command)
-
-      expect(logger).to have_received(:write).with("\n").twice
-      expect(logger).to have_received(:write_no_newline).with("hello wo")
-      expect(logger).to have_received(:write_no_newline).with("rld\\ntod")
-      expect(logger).to have_received(:write_no_newline).with("o roto")
-    end
-
-    context 'when the execution was succsessfull' do
-      it 'prints a new line' do
-        runner.execute(command)
-        expect(logger).to have_received(:write).twice.with(/\n/)
-      end
-    end
-
-    context 'on failure' do
-      let(:expected_command) { "#{command} 2> #{config.error_log_path}" }
-
-      before do
-        allow(Open3).to(
-          receive(:popen3)
-          .with(expected_command)
-          .and_call_original
-        )
-      end
-
-      context 'when the execution failed' do
-        let(:command) { 'sh -c \'echo ROTO >/dev/stderr && false\'' }
-
-        it 'raises a PerconaMigrator::Error' do
-          expect { runner.execute(command) }
-            .to raise_exception(PerconaMigrator::Error, "ROTO\n")
-        end
-      end
-
-      context 'when the command was signaled' do
-        let(:command) { 'kill -9 $$' }
-
-        it 'raises a SignalError specifying the status' do
-          expect { runner.execute(command) }
-            .to raise_exception(PerconaMigrator::SignalError)
-        end
-      end
-
-      context 'when pt-online-schema-change is not installed' do
-        let(:command) { 'whatevarrr666' }
-
-        it 'raises a detailed CommandNotFoundError' do
-          expect { runner.execute(command) }
-            .to raise_exception(
-              PerconaMigrator::CommandNotFoundError,
-              /Please install pt-online-schema-change/
-            )
-        end
-      end
     end
   end
 end
